scheme: rearrange jobs association list
The association list created by guile-json from the test files is not
too convenient to use. for one, it creates arrays rather than lists.
Furthermore, the test files don't specify lists where lists are empty,
namely, empty lists of "after" and "before".

Recreate the jobs in a format that is easier to work with.