scheme: reimplement alist-equal?
The implementation of alist-equal? was simply wrong and produced wrong
results. Namely, it did not traverse deep enough to handle inconsistent
ordering of elements in lists.

The function has been reimplemented. Sadly, it is no longer generalized,
but instead it is specialized to compare two alists comprised of jobs.
Thus, the function has been renamed as jobs-alist-equal?